requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.8
Django==5.0.6
django-ckeditor-5==0.2.15
django-debug-toolbar==4.4.6
django-js-asset==2.2.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-rosetta==0.10.1
filelock==3.16.1
fontawesomefree==6.0.0
idna==3.10
packaging==24.1
pillow==10.3.0
platformdirs==4.3.6
pluggy==1.5.0
polib==1.2.0
Pygments==2.19.1
pyproject-api==1.8.0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.6
